more securing..

+++ content/delete_document.php
@@ -1,13 +1,13 @@
 <?php
     include("../top.php");
-    $uid = $_SESSION['session_userid'];
-    $user = new User($uid);
-    $rPageID = $_POST['rPageID'];
 
-    $doc = new Document($rPageID);
-    $title = $doc->getTitle();
-    $parent_id = $doc->get('parent_id');
-    $doc->delete();
-    chatline("deletes document $title.");
-    print $parent_id;
+    if($access == 0){
+        $rPageID = $_POST['rPageID'];
+        $doc = new Document($rPageID);
+        $title = $doc->getTitle();
+        $parent_id = $doc->get('parent_id');
+        $doc->delete();
+        chatline("deletes document $title.");
+        print $parent_id;
+    }
